--- packages/compiler/README.md
@@ -1,15 +1,0 @@
-# ovm-compiler
-
-## Usage
-
-- [ovm-compiler](https://github.com/cryptoeconomicslab/ovm-compiler/tree/master/packages/compiler)
-
-## Online Demo
-
-https://ovm-compiler.netlify.com
-
-## Test
-
-```
-npm test
-```
