--- packages/compiler/package.json
@@ -1,57 +1,0 @@
-{
-  "name": "@cryptoeconomicslab/ovm-compiler",
-  "version": "0.2.3",
-  "description": "OVM compiler",
-  "author": {
-    "name": "Cryptoeconomics Lab",
-    "url": "https://www.cryptoeconomicslab.com/"
-  },
-  "engines": {
-    "node": ">=10"
-  },
-  "license": "Apache-2.0",
-  "main": "lib/index.js",
-  "directories": {
-    "lib": "lib",
-    "test": "__tests__"
-  },
-  "files": [
-    "lib"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "tsc --build",
-    "build:check": "tsc --noEmit",
-    "test": "jest --passWithNoTests",
-    "lint": "eslint --ext ts,tsx src",
-    "doc": "typedoc --out docs src",
-    "clean": "rm -rf lib docs"
-  },
-  "jest": {
-    "transform": {
-      "^.+\\.tsx?$": "ts-jest"
-    },
-    "testMatch": [
-      "**/__tests__/**/?(*.)+(spec|test).(ts|js)",
-      "**/?(*.)+(spec|test).ts"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "ts"
-    ],
-    "globals": {
-      "ts-jest": {
-        "tsConfig": "tsconfig.json"
-      }
-    }
-  },
-  "dependencies": {
-    "@cryptoeconomicslab/ovm-ethereum-generator": "^0.2.3",
-    "@cryptoeconomicslab/ovm-parser": "^0.2.3",
-    "@cryptoeconomicslab/ovm-solidity-generator": "^0.2.3",
-    "@cryptoeconomicslab/ovm-transpiler": "^0.2.3",
-    "tslib": "^1.10.0"
-  }
-}
