--- packages/compiler/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./lib"
-  },
-  "include": ["./src"]
-}
